remove viewer class and make GuessTheNatoWord the viewer
