+++ src/main/java/com/nato/GuessTheNatoWordController.java
@@ -1,0 +1,7 @@
+package com.nato;
+
+public class GuessTheNatoWordController {
+    GuessTheNatoWordController() {
+
+    }
+}
