+++ src/main/java/com/nato/GuessTheNatoWordViewer.java
@@ -3,16 +3,16 @@ package com.nato;
 import javax.swing.*;
 import java.awt.*;
 
-public class GuessTheNatoWordApp {
-    private final JFrame viewer;
+public class GuessTheNatoWordViewer extends JFrame {
     private final JPanel parentPanel;
     private final IntroPanel introPanel;
 
-    public GuessTheNatoWordApp(String INTRO_MESSAGE, JFrame viewer) {
-        this.viewer = viewer;
+
+    public GuessTheNatoWordViewer(String INTRO_MESSAGE) {
         this.parentPanel = new JPanel();
         this.introPanel = new IntroPanel(INTRO_MESSAGE);
     }
+
 
     public void injectNextPanel(JPanel nextPanel) {
         parentPanel.add(nextPanel);
@@ -23,14 +23,29 @@ public class GuessTheNatoWordApp {
         parentPanel.repaint();
     }
 
+
     public void setParentPanel() {
         parentPanel.setLayout(new GridBagLayout());
-        viewer.add(parentPanel);
+        this.add(parentPanel);
     }
+
 
     public void setPanels() {
         introPanel.set();
         setParentPanel();
+    }
+
+    public void setFrame() {
+        this.setExtendedState(this.getExtendedState() | JFrame.MAXIMIZED_BOTH); // set frame to maximized
+        this.setSize(1000, 680); // size in case the window is unmaximized
+        this.setLocationRelativeTo(null); // center the unmaximized frame in the screen
+        this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        this.setVisible(true);
+    }
+
+    public void set() {
+        setPanels();
+        setFrame();
     }
 
     public void waitForNSeconds(int seconds) {
@@ -50,7 +65,7 @@ public class GuessTheNatoWordApp {
     }
 
     public void play() {
-        setPanels();
+        set();
         showIntro();
     }
 }
